==> scripts/s01_read_data_feb2016.Rmd <==
---
title: "Read data, wecare feb2016"
author: "Alexey Larionov"
output: html_document
---

started: 01Mar2016
last updated: 12May2016

# start_section

```{r start_section}

# Time stamp
Sys.time()

# Clean-up
rm(list=ls())
graphics.off()

# Set root working folder
library(knitr)

opts_knit$set(root.dir = "/scratch/medgen/scripts/rscripts_05.16")
#setwd("/scratch/medgen/scripts/rscripts_05.16")

#opts_knit$set(root.dir = "C:\\Users\\larion01\\Documents\\GitHub\\R1")
#setwd("C:\\Users\\larion01\\Documents\\GitHub\\R1")

# Required libraries
library(ggplot2)

```

# link_WES_samples_ids_with_GWAS_ids_and_annotations

```{r link_WES_samples_ids_with_GWAS_ids_and_annotations}

# Make text file linking samples IDs in GWAs and WES data
#load("data/samples_AL09Nov2015.RData")
#write.table(samples.ids, file="data/samples_ids.txt", quote="", sep="\t")

# Check content of full gwas data
#load("data/gwas_AL10Jul2014.RData")

```

# read_data

```{r read_data}

covar.df <- read.table("data/covar.txt", sep="\t", header=TRUE, quote="")
demographics.df <- read.table("data/WECARE.Exome.DemographicVariables.txt", sep="\t", header=TRUE, quote="")

samples.df <- read.table("data/samples_ids.txt", sep="\t", header=TRUE, quote="")

vv.df <- read.table("data/wecare_feb2016_std_filter_VV_biallelic.txt", header=TRUE, row.names=1, sep="\t", quote="")

gt.df <- read.table("data/wecare_feb2016_std_filter_GT_biallelic_add.txt", header=TRUE, row.names=1, sep="\t", quote="")
gq.df <- read.table("data/wecare_feb2016_std_filter_GQ_biallelic.txt", header=TRUE, row.names=1, sep="\t", quote="")
dp.df <- read.table("data/wecare_feb2016_std_filter_DP_biallelic.txt", header=TRUE, row.names=1, sep="\t", quote="")
ad.df <- read.table("data/wecare_feb2016_std_filter_AD_biallelic.txt", header=TRUE, row.names=1, sep="\t", quote="")

```

# check_data

```{r check_data}

dim(covar.df)
str(covar.df)
covar.df[1:5,1:5]

dim(demographics.df)
str(demographics.df)
demographics.df[1:5,1:5]

dim(samples.df)
str(samples.df)
samples.df[1:5,]

dim(vv.df)
str(vv.df)
vv.df[1:5,1:5]

dim(gt.df)
str(gt.df, list.len=5)
gt.df[1:5,1:5]

dim(gq.df)
str(gq.df, list.len=5)
gq.df[1:5,1:5]

dim(dp.df)
str(dp.df, list.len=5)
dp.df[1:5,1:5]

dim(ad.df)
str(ad.df, list.len=5)
ad.df[1:15,1:5]

```

# explore_age
Inclusion criteria: age at 1st Ds < 55

```{r explore_age}

ggplot(covar.df, aes(sub_dx_age)) +
  geom_histogram(binwidth = 5, colour="black", alpha=0.25) + 
  aes(fill=as.factor(cc)) +
  ggtitle("Age at the 1st diagnosis in cases and controls")

```

# explore_time_to_2nd_ca
Inclusion criteria: 2nd Ds should be 1yr+ later than the 1st Ds
There are two fields that look appropriate ...

```{r explore_time_to_2nd_ca}

ggplot(covar.df, aes(offset)) +
  geom_histogram(breaks=c(0:8), right=FALSE, colour="black", alpha=0.25) + 
  aes(fill=as.factor(cc)) +
  ggtitle("covar: offset")

ggplot(demographics.df[c(-59, -502,-503,-504,-505),], aes(rstime)) +
  geom_histogram(breaks=c(0:8), right=FALSE, colour="black", alpha=0.25) + 
  aes(fill=as.factor(cc.y)) +
  ggtitle("demographics: rstime")

```

# split_ad_table

```{r split_ad_table}

# Convert source data frame to matrix
ad.mx <- as.matrix(ad.df)
rm(ad.df)

# Prepare matrices for output
ref.mx <- ad.mx
alt.mx <- ad.mx

# Split ad into two tables (long step)
for (row in 1:nrow(ad.mx)){
  for (col in 1:ncol(ad.mx)){
    ref.alt <- strsplit(ad.mx[row,col],",")[[1]]
    ref.mx[row,col] <- ref.alt[1]
    alt.mx[row,col] <- ref.alt[2]
  }
}

# Clean-up
rm(row, col, ref.alt)

# Explore results
dim(ad.mx)
dim(ref.mx)
dim(alt.mx)

ad.mx[1:15,1:5]
ref.mx[1:15,1:5]
alt.mx[1:15,1:5]

# Clean-up
rm(ad.mx)

```

# save_data

```{r save_data}

save.image(file="data/s01_read_data_feb2016.RData")

```

# final_section

```{r final_section}

ls()
sessionInfo()
Sys.time()

```
