==> scripts/s02_filter_mask_feb2016.Rmd <==
---
title: "Filter and mask variants, wecare feb2016"
author: "Alexey Larionov"
output: html_document
---

started: Mar2016
last updated: 29Mar2016

# start_section

```{r start_section}

# Time stamp
Sys.time()

# Clean-up
rm(list=ls())
graphics.off()

# Set root working folder
library(knitr)
opts_knit$set(root.dir = "/scratch/medgen/users/alexey/wecare_feb2016/wecare_feb2016_r")
#setwd("/scratch/medgen/users/alexey/wecare_feb2016/wecare_feb2016_r")

```

# load_and_check_data

```{r load_and_check_data}

load(file="data/s01_read_data_feb2016.RData")
ls()

dim(covar.df)
colnames(covar.df)
str(covar.df)
covar.df[1:5,1:5]

dim(samples.df)
colnames(samples.df)
str(samples.df)
samples.df[1:5,]

dim(vv.df)
colnames(vv.df)
str(vv.df)
vv.df[1:5,1:3]

dim(gt.df)
gt.df[1:5,1:3]

dim(gq.df)
gq.df[1:5,1:3]

dim(dp.df)
dp.df[1:5,1:3]

# Check consistence of rownames
sum(rownames(gt.df) != rownames(gq.df), na.rm=TRUE)
sum(rownames(gt.df) != rownames(dp.df), na.rm=TRUE)
sum(rownames(gt.df) != rownames(vv.df), na.rm=TRUE)

# Check consistence of colnames
sum(substr(colnames(gt.df),0,6) != substr(colnames(gq.df),0,6), na.rm=TRUE)
sum(substr(colnames(gt.df),0,6) != substr(colnames(dp.df),0,6), na.rm=TRUE)

```

# filter_variants

```{r filter_variants}

# Genoptypes NA rate before variants filtering
sum(is.na(gt.df))/(dim(gt.df)[1]*dim(gt.df)[2]) # ~6%

vars.pf <- vv.df$FILTER == "PASS"
sum(vars.pf)

gt.pf.df <- gt.df[vars.pf,]
dim(gt.pf.df)
rm(gt.df)

gq.pf.df <- gq.df[vars.pf,]
dim(gq.pf.df)
rm(gq.df)

dp.pf.df <- dp.df[vars.pf,]
dim(dp.pf.df)
rm(dp.df)

vv.pf.df <- vv.df[vars.pf,]
dim(vv.pf.df)
rm(vv.df)

rm(vars.pf)

# Genoptypes NA rate after variants filtering
sum(is.na(gt.pf.df))/(dim(gt.pf.df)[1]*dim(gt.pf.df)[2]) # <1%

```

# mask_genotypes

```{r mask_genotypes}

# Converet to matrices
gt.pf.mx <- as.matrix(gt.pf.df)
gq.pf.mx <- as.matrix(gq.pf.df)
dp.pf.mx <- as.matrix(dp.pf.df)
rm(gt.pf.df, gq.pf.df, dp.pf.df)

# Types of genotypes to be excluded by masking
sum(gq.pf.mx < 20, na.rm=TRUE)/(dim(gt.pf.mx)[1]*dim(gt.pf.mx)[2]) # ~5%
sum(dp.pf.mx < 10, na.rm=TRUE)/(dim(dp.pf.mx)[1]*dim(dp.pf.mx)[2]) # ~7%

# Mask low quality genotypes
gt.pf.msk.mx <- gt.pf.mx
NA -> gt.pf.msk.mx[ gq.pf.mx <20 ]
NA -> gt.pf.msk.mx[ dp.pf.mx <10 ]
rm(gt.pf.mx, gq.pf.mx, dp.pf.mx)

# Genoptypes NA rate after filtering and masking
sum(is.na(gt.pf.msk.mx))/(dim(gt.pf.msk.mx)[1]*dim(gt.pf.msk.mx)[2]) # ~7%
dim(gt.pf.msk.mx)

```

# save_data

```{r save_data}

save.image(file="data/s02_filter_mask_feb2016.RData")

```

# final_section

```{r final_section}

sessionInfo()
Sys.time()

```
