==> scripts/s07_read_1k.Rmd <==
---
title: "Read 1k data"
author: "AL"
output: html_document
---

started: Mar2016
last_updated:26Mar2016

# start_section

```{r start_section}

# Start time
Sys.time()

# Clean-up
rm(list=ls())
graphics.off()

# Set root working folder
library(knitr)
opts_knit$set(root.dir = "/scratch/medgen/users/alexey/wecare_feb2016/wecare_feb2016_r")
#setwd("/scratch/medgen/users/alexey/wecare_feb2016/wecare_feb2016_r")

```

# read_data

```{r read_data}

# Source txt data have been copied from from resources forlder to local data folder:
#cp /scratch/medgen/resources/phase3_1k_release20130502/wes_vep/nextera_exome_ip10_biallelic_nonsv_1k_vep_vcf.txt 

ph3_1k.df <- read.table("data/nextera_exome_ip10_biallelic_nonsv_1k_vep_vcf.txt", sep="\t", header=TRUE, quote="")

```

# check_data

```{r check_data}

dim(ph3_1k.df)
colnames(ph3_1k.df)
str(ph3_1k.df)
ph3_1k.df[1:5, 1:7]

```

# Notes about MAFs cleaning

Inspection of the imported data was done using str() and summary() of selected fields. 
Particular attention was paid to the type of columns, which were supposed to be numeric. 

The inspection showed that MAF fields generated by VEP were not numeric. 
Further focused exploration of these fields showed that there was a small number (<1%) 
of variants where VEP reported confusing MAFs, which looked like MAFs for multiple alleles, 
despite all variants were biallelic by GATK. These MAFs could not be converted to 
numeric, so they were removed (see below).

The MAF data are not used in downstream analysis. However, 
the cleaning was done just as a matter of keeping data tidy.  

No cleaning was done for other numeric fields: cDNA_position, CDS_position 
and Protein_position. None of the fields will be used in the downstream analysis either. 

# cleaning_gmaf

``` {r cleaning_gmaf}

gmaf_str <- as.vector(ph3_1k.df$GMAF)
sum(is.na(gmaf_str))

gmaf_num <- as.numeric(gmaf_str)
sum(is.na(gmaf_num))

odd_gmaf <- is.na(gmaf_num) & ! is.na(gmaf_str)
sum(odd_gmaf)
sum(odd_gmaf)/length(odd_gmaf)

odd_gmaf.df <- ph3_1k.df[odd_gmaf,]
odd_gmaf.df[c(1:10),c("CHROM","POS","REF","ALT","AF","GMA","GMAF")]

NA -> ph3_1k.df$GMA[odd_gmaf]
gmaf_num -> ph3_1k.df$GMAF
summary(ph3_1k.df$GMAF)

rm(gmaf_str, gmaf_num, odd_gmaf, odd_gmaf.df)

```

# cleaning_afr_maf

```{r cleaning_afr_maf}

afr_maf_str <- as.vector(ph3_1k.df$AFR_MAF)
sum(is.na(afr_maf_str))

afr_maf_num <- as.numeric(afr_maf_str)
sum(is.na(afr_maf_num))

odd_afr_maf <- is.na(afr_maf_num) & ! is.na(afr_maf_str)
sum(odd_afr_maf)
sum(odd_afr_maf)/length(odd_afr_maf)

odd_afr_maf.df <- ph3_1k.df[odd_afr_maf,]
odd_afr_maf.df[c(1:10),c("CHROM","POS","REF","ALT","AFR_AF","AFR_MA","AFR_MAF")]

NA -> ph3_1k.df$AFR_MA[odd_afr_maf]
afr_maf_num -> ph3_1k.df$AFR_MAF
summary(ph3_1k.df$AFR_MAF)

rm(afr_maf_str, afr_maf_num, odd_afr_maf, odd_afr_maf.df)

```

# cleaning_eur_maf

```{r cleaning_eur_maf}

eur_maf_str <- as.vector(ph3_1k.df$EUR_MAF)
sum(is.na(eur_maf_str))

eur_maf_num <- as.numeric(eur_maf_str)
sum(is.na(eur_maf_num))

odd_eur_maf <- is.na(eur_maf_num) & ! is.na(eur_maf_str)
sum(odd_eur_maf)
sum(odd_eur_maf)/length(odd_eur_maf)

odd_eur_maf.df <- ph3_1k.df[odd_eur_maf,]
odd_eur_maf.df[c(1:10),c("CHROM","POS","REF","ALT","EUR_AF","EUR_MA","EUR_MAF")]

NA -> ph3_1k.df$EUR_MA[odd_eur_maf]
eur_maf_num -> ph3_1k.df$EUR_MAF
summary(ph3_1k.df$EUR_MAF)

rm(eur_maf_str, eur_maf_num, odd_eur_maf, odd_eur_maf.df)

```

# cleaning_asn_maf

```{r cleaning_asn_maf}

# No data on ASN_MAF

```

# cleaning_eas_maf

```{r cleaning_eas_maf}

eas_maf_str <- as.vector(ph3_1k.df$EAS_MAF)
sum(is.na(eas_maf_str))

eas_maf_num <- as.numeric(eas_maf_str)
sum(is.na(eas_maf_num))

odd_eas_maf <- is.na(eas_maf_num) & ! is.na(eas_maf_str)
sum(odd_eas_maf)
sum(odd_eas_maf)/length(odd_eas_maf)

odd_eas_maf.df <- ph3_1k.df[odd_eas_maf,]
odd_eas_maf.df[c(1:10),c("CHROM","POS","REF","ALT","EAS_AF","EAS_MA","EAS_MAF")]

NA -> ph3_1k.df$EAS_MA[odd_eas_maf]
eas_maf_num -> ph3_1k.df$EAS_MAF
summary(ph3_1k.df$EAS_MAF)

rm(eas_maf_str, eas_maf_num, odd_eas_maf, odd_eas_maf.df)

```

# cleaning_sas_maf

```{r cleaning_sas_maf}

sas_maf_str <- as.vector(ph3_1k.df$SAS_MAF)
sum(is.na(sas_maf_str))

sas_maf_num <- as.numeric(sas_maf_str)
sum(is.na(sas_maf_num))

odd_sas_maf <- is.na(sas_maf_num) & ! is.na(sas_maf_str)
sum(odd_sas_maf)
sum(odd_sas_maf)/length(odd_sas_maf)

odd_sas_maf.df <- ph3_1k.df[odd_sas_maf,]
odd_sas_maf.df[c(1:10),c("CHROM","POS","REF","ALT","EAS_AF","SAS_MA","SAS_MAF")]

NA -> ph3_1k.df$SAS_MA[odd_sas_maf]
sas_maf_num -> ph3_1k.df$SAS_MAF
summary(ph3_1k.df$SAS_MAF)

rm(sas_maf_str, sas_maf_num, odd_sas_maf, odd_sas_maf.df)

```

# cleaning_aa_maf

```{r cleaning_aa_maf}

aa_maf_str <- as.vector(ph3_1k.df$AA_MAF)
sum(is.na(aa_maf_str))

aa_maf_num <- as.numeric(aa_maf_str)
sum(is.na(aa_maf_num))

odd_aa_maf <- is.na(aa_maf_num) & ! is.na(aa_maf_str)
sum(odd_aa_maf)
sum(odd_aa_maf)/length(odd_aa_maf)

odd_aa_maf.df <- ph3_1k.df[odd_aa_maf,]
odd_aa_maf.df[c(1:10),c("CHROM","POS","REF","ALT","AF","AA_MA","AA_MAF")]

NA -> ph3_1k.df$AA_MA[odd_aa_maf]
aa_maf_num -> ph3_1k.df$AA_MAF
summary(ph3_1k.df$AA_MAF)

rm(aa_maf_str, aa_maf_num, odd_aa_maf, odd_aa_maf.df)

```

# cleaning_ea_maf

```{r cleaning_ea_maf}

ea_maf_str <- as.vector(ph3_1k.df$EA_MAF)
sum(is.na(ea_maf_str))

ea_maf_num <- as.numeric(ea_maf_str)
sum(is.na(ea_maf_num))

odd_ea_maf <- is.na(ea_maf_num) & ! is.na(ea_maf_str)
sum(odd_ea_maf)
sum(odd_ea_maf)/length(odd_ea_maf)

odd_ea_maf.df <- ph3_1k.df[odd_ea_maf,]
odd_ea_maf.df[c(1:10),c("CHROM","POS","REF","ALT","AF","EA_MA","EA_MAF")]

NA -> ph3_1k.df$EA_MA[odd_ea_maf]
ea_maf_num -> ph3_1k.df$EA_MAF
summary(ph3_1k.df$EA_MAF)

rm(ea_maf_str, ea_maf_num, odd_ea_maf, odd_ea_maf.df)

```

# check_cleaned_data

```{r check_cleaned_data}

str(ph3_1k.df)

```

# save_data

```{r save_data}

save.image(file="data/s07_read_1k.RData")

```

# final_section

```{r final_section}

sessionInfo()
Sys.time()
  
```
